+++ src/app/components/employee/employee-list/employee-list.component.ts
@@ -2,6 +2,7 @@ import { ToastrService } from 'ngx-toastr';
 import { EmployeeService } from './../../../services/employee.service';
 import { Component, OnInit } from '@angular/core';
 import { IEmployeeGetAllModel } from 'src/app/models/response/employee/employee-getall-model';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-employee-list',
@@ -17,23 +18,55 @@ export class EmployeeListComponent implements OnInit {
   ) {}
 
   ngOnInit(): void {
-    console.log("sfsdfsdf");
-    
+    console.log('sfsdfsdf');
+
     this.getEmployees();
   }
 
   getEmployees() {
     this.employeeService.getEmployees().subscribe((data) => {
-      console.log(data);
-      
       this.employees = data;
     });
   }
 
   deleteToEmployee(id: number) {
-    this.employeeService.deleteToEmployee(id).subscribe(() => {
-      this.toastrService.success('Silme İşlemi Gerçekleşti', 'Tebrikler (:');
-      window.location.reload();
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: 'btn btn-success',
+        cancelButton: 'btn btn-danger',
+      },
+      buttonsStyling: false,
     });
+
+    swalWithBootstrapButtons
+      .fire({
+        title: 'Emin misiniz?',
+        text: 'Bu işlem geri alınamaz!',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Evet!',
+        cancelButtonText: 'Hayır',
+        reverseButtons: true,
+      })
+      .then((result) => {
+        if (result.isConfirmed) {
+          this.employeeService.deleteToEmployee(id).subscribe(() => {
+            swalWithBootstrapButtons.fire(
+              'Silindi!',
+              'İstediğiniz veri silme işlemi tamamlandı.',
+              'success'
+            );
+          }),
+            setTimeout(() => {
+              window.location.reload();
+            }, 3000);
+        } else if (result.dismiss === Swal.DismissReason.cancel) {
+          swalWithBootstrapButtons.fire(
+            'İptal edildi',
+            'Veriniz hala güvende :)',
+            'error'
+          );
+        }
+      });
   }
 }
